server: Move CubebServer to server submodule, clean up top-level lib.rs.
